requirements.txt
langchain
openai
streamlit
python-dotenv
wikipedia
numexpr
youtube-transcript-api
tiktoken
faiss-cpu
